sos - refactored engine sos plugin
engine sos plugin used two flags: vdsmlogs and prefix.
Both flags were implemented with behaviors not acceptable by sos
developers:
- vdsmlogs is not a boolean flag: it was used for specifying a
  path containing logs collected by sosreport runs on other hosts;
  if required the path should have been determined without allowing
  the user to specify a generic path like '/'
- prefix is not a boolean flag and was used for changing the name
  of the report tarball. The name was changed altering the sosreport
  policy instead of creating a new policy.

Both these 2 flag are meaningless if you run sosreport -a -o engine
because -a set both the flags to True and because running sosreport
you can't collect hypervisors sosreport in vdsmlogs and you want
a standard sosreport.

I've refactored the engine sos plugin removing both the flags.
It's engine-log-collector that will take care to rename the
tarball with the desired prefix and it's engine-log-collector
that will take care of adding the vdsmlogs to the archive.

PREVIOUS:
- collect and archive postgres data in a scratch dir
- collect and archive hypervisors data in a scratch dir
- collect and archive engine data (including the above in /tmp)

NOW:
- collect engine data
- collect and archive postgres data in log-collector-data
- collect and archive hypervisors data in log-collector-data
- archive all the above in a single tarball.

Change-Id: Ib226ec23b98450a03ebd2a64043a911645a62235
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>