Use max=1000 option return results when calling the oVirt API.
When calling the API the LC does not provide a max value, limiting
the results to 100 by default. For environments with 100+ DCs,
Clusters or hosts this leads to incomplete LCs being collected.

This change introduces the max=1000 option to each .list() call to allow
a greater number of results to be returned.

Change-Id: I0c9f61938df1626b1d81e4c0320b5bca10614f34
Bug-Url: https://bugzilla.redhat.com/1014379
Signed-off-by: Lee Yarwood <lyarwood@redhat.com>
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>