HyperVisorData: Run ssh with -n
The fix to https://bugzilla.redhat.com/1010472 was to add '-t' to
provide a controlling tty to the remote ausearch. This caused the
tty to be garbled under certain conditions.

A comment in that bug mentions that this is due to ausearch expecting a
log file path on stdin if stdin is a pipe. So this change makes ssh run
with '-n' thus not providing an stdin to the remote command.

Change-Id: Id3e65b5564eb34e72e57f824fb86f9c50acfa1c5
Bug-Url: https://bugzilla.redhat.com/1058744
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>
Signed-off-by: Yedidyah Bar David <didi@redhat.com>