importDumpIntoNewDb: initdb - do not redirect standard output error
Currently, if users execute initdb with root or other users
not allowed to use postgres, the error message will be redirected
to initdb.log and won't show in the console. This patch removes
the redirect and the error now show in the console to users as
soon an it appears.

Change-Id: I22a2336f3b52cc11d9fd2a4027dddfa03d0ba169
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>