CollectorBase: fix list index out of range
Re-ordered tests on an list avoiding an index out
of range exception if the sosreport filename is not
in the output of sosreport call.

The IndexError was used to detect an error in sosreport call
instead of properly handling a parsing error. Added exception
handling for this cases and made it a critical error since
this means the underlying sosreport is having troubles.

Change-Id: I40235242bf4eabd4b1cd9599eb45cb124e182df3
Bug-Url: https://bugzilla.redhat.com/1460734
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>